Fail closed on malformed cloud benchmark usage

+++ crates/goose-provider-types/src/conversation/message.rs
@@ -1085,20 +1085,20 @@ impl Message {
 #[derive(Debug, Clone, Default, Serialize, Deserialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct TokenState {
-    pub input_tokens: i32,
-    pub output_tokens: i32,
-    pub total_tokens: i32,
+    pub input_tokens: i64,
+    pub output_tokens: i64,
+    pub total_tokens: i64,
     #[serde(default)]
-    pub cache_read_tokens: i32,
+    pub cache_read_tokens: i64,
     #[serde(default)]
-    pub cache_write_tokens: i32,
-    pub accumulated_input_tokens: i32,
-    pub accumulated_output_tokens: i32,
-    pub accumulated_total_tokens: i32,
+    pub cache_write_tokens: i64,
+    pub accumulated_input_tokens: i64,
+    pub accumulated_output_tokens: i64,
+    pub accumulated_total_tokens: i64,
     #[serde(default)]
-    pub accumulated_cache_read_tokens: i32,
+    pub accumulated_cache_read_tokens: i64,
     #[serde(default)]
-    pub accumulated_cache_write_tokens: i32,
+    pub accumulated_cache_write_tokens: i64,
     pub accumulated_cost: Option<f64>,
 }
