+++ crates/goose/src/agents/platform_extensions/apps.rs
@@ -282,7 +282,7 @@ impl AppsManagerClient {
         .map_err(|e| format!("LLM call failed: {}", e))?;
 
         if let (Some(output), Some(max)) = (usage.usage.output_tokens, model_config.max_tokens) {
-            if output >= max {
+            if output >= i64::from(max) {
                 return Err("App content generation was truncated because the response hit the token limit. Try simplifying your app description.".to_string());
             }
         }
@@ -327,7 +327,7 @@ impl AppsManagerClient {
         .map_err(|e| format!("LLM call failed: {}", e))?;
 
         if let (Some(output), Some(max)) = (usage.usage.output_tokens, model_config.max_tokens) {
-            if output >= max {
+            if output >= i64::from(max) {
                 return Err("App content update was truncated because the response hit the token limit. Try requesting smaller changes.".to_string());
             }
         }
